lesson_1_fix

+++ Lesson_1/area_of_a_triangle.rb
@@ -1,9 +1,9 @@
 puts "Введите основание треугольника"
-base = gets.to_i
+base = gets.to_f
 
 puts "Введите высоту треугольника"
-height = gets.to_i
+height = gets.to_f
 
-area = 0.5*base*height
+area = 0.5 * base * height
 
 puts "Площадь треугольника #{area.round(2)}"
